refactor(plots): move module to package
- blank module retained for diff
